refactor(auth): extract API key CRUD routes from auth_routes() into ServerAppBuilder
API key CRUD routes (/v1/auth/api-keys) are auth-provider-agnostic but
were bundled inside AuthBackend::auth_routes(). Custom AuthBackend
implementations (e.g. PropelAuth) had to reimplement auth_routes()
entirely and silently lost API key endpoints.

Extract api_key_routes() as a standalone function (same pattern as
cli_auth_routes()) and mount it directly in ServerAppBuilder. Add
on_api_key_deleted() to AuthBackend trait for cache invalidation.

Closes EVE-272
