+++ crates/server/src/auth/backend.rs
@@ -21,9 +21,16 @@ pub trait AuthBackend: Send + Sync + 'static {
     /// Validate an API key and return the authenticated user.
     async fn validate_api_key(&self, key: &str) -> Result<AuthUser, AuthError>;
 
-    /// Return auth-specific HTTP routes (login, register, OAuth callbacks, API key CRUD).
+    /// Return auth-specific HTTP routes (login, register, OAuth callbacks).
     /// Returns `None` if auth is handled externally (e.g., PropelAuth hosted UI).
+    ///
+    /// API key CRUD routes are mounted separately by `ServerAppBuilder` via
+    /// `api_key_routes()` — they are auth-provider-agnostic.
     fn auth_routes(&self) -> Option<Router>;
+
+    /// Called when an API key is deleted, so backends can invalidate caches.
+    /// Default: no-op (backends without caching don't need to override).
+    fn on_api_key_deleted(&self) {}
 
     /// Return root-level public routes that must not be nested under the API prefix.
     ///
